+++ src/lib/Icons/IconClose.svelte
