+++ src/lib/Icons/IconLibrary.docs.mdx
@@ -4,7 +4,7 @@ import * as CloseStory from "./IconClose.stories.svelte";
 import * as SearchStory from "./IconSearch.stories.svelte";
 import { Meta, Title, Canvas, Story, Controls, IconGallery, IconItem } from "@storybook/blocks";
 
-<Meta title="Components/IconLibrary" of={PlusStories} />
+<Meta title="Components/Icons" of={PlusStories} />
 
 # Icons
