+++ src/lib/Icons/IconMinus.svelte
