+++ src/lib/Icons/IconPlus.svelte
