+++ src/lib/Icons/IconSearch.svelte
