+++ src/lib/index.js
@@ -3,10 +3,10 @@ export { default as BasicDropdown } from "./BasicDropdown/BasicDropdown.svelte";
 export { default as Block } from "./Block/Block.svelte";
 export { default as DatawrapperIframe } from "./DatawrapperIframe/DatawrapperIframe.svelte";
 export { default as HeadingAlt } from "./HeadingAlt/HeadingAlt.svelte";
-export { default as IconPlus } from "./IconLibrary/IconPlus.svelte";
-export { default as IconMinus } from "./IconLibrary/IconMinus.svelte";
-export { default as IconClose } from "./IconLibrary/IconClose.svelte";
-export { default as IconSearch } from "./IconLibrary/IconSearch.svelte";
+export { default as IconPlus } from "./Icons/IconPlus.svelte";
+export { default as IconMinus } from "./Icons/IconMinus.svelte";
+export { default as IconClose } from "./Icons/IconClose.svelte";
+export { default as IconSearch } from "./Icons/IconSearch.svelte";
 export { default as LogoTPCBadge } from "./LogoTPCBadge/LogoTPCBadge.svelte";
 export { default as LogoUrbanBadge } from "./LogoUrbanBadge/LogoUrbanBadge.svelte";
 export { default as LogoUrbanAnimated } from "./LogoUrbanAnimated/LogoUrbanAnimated.svelte";
